feat(config): allow custom ollama path
