+++ config/config.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"runtime"
 	"time"
 
 	"github.com/fsnotify/fsnotify"
@@ -15,6 +16,7 @@ type Config struct {
 	Columns           []string `mapstructure:"columns"`
 	OllamaAPIKey      string   `mapstructure:"ollama_api_key"`
 	OllamaAPIURL      string   `mapstructure:"ollama_api_url"`
+	OllamaModelsDir   string   `mapstructure:"ollama_models_dir"`
 	LMStudioFilePaths string   `mapstructure:"lm_studio_file_paths"`
 	LogLevel          string   `mapstructure:"log_level"`
 	LogFilePath       string   `mapstructure:"log_file_path"`
@@ -29,12 +31,25 @@ var defaultConfig = Config{
 	Columns:           []string{"Name", "Size", "Quant", "Family", "Modified", "ID"},
 	OllamaAPIKey:      "",
 	OllamaAPIURL:      getAPIUrl(),
+	OllamaModelsDir:   GetOllamaModelDir(),
 	LMStudioFilePaths: "",
 	LogLevel:          "info",
 	SortOrder:         "modified",
 	StripString:       "",
 	Editor:            "/usr/bin/vim",
 	DockerContainer:   "",
+}
+
+// GetOllamaModelDir returns the default Ollama models directory for the current OS
+func GetOllamaModelDir() string {
+	homeDir := utils.GetHomeDir()
+	if runtime.GOOS == "darwin" {
+		return filepath.Join(homeDir, ".ollama", "models")
+	} else if runtime.GOOS == "linux" {
+		return "/usr/share/ollama/models"
+	}
+	// Add Windows path if needed
+	return filepath.Join(homeDir, ".ollama", "models")
 }
 
 // getAPIUrl determines the API URL based on environment variables.
@@ -108,6 +123,7 @@ func LoadConfig() (Config, error) {
 	viper.SetDefault("columns", defaultConfig.Columns)
 	viper.SetDefault("ollama_api_key", defaultConfig.OllamaAPIKey)
 	viper.SetDefault("ollama_api_url", defaultConfig.OllamaAPIURL)
+	viper.SetDefault("ollama_models_dir", defaultConfig.OllamaModelsDir)
 	viper.SetDefault("lm_studio_file_paths", defaultConfig.LMStudioFilePaths)
 	viper.SetDefault("log_level", defaultConfig.LogLevel)
 	viper.SetDefault("log_file_path", defaultConfig.LogFilePath)
@@ -119,6 +135,7 @@ func LoadConfig() (Config, error) {
 	config.Columns = viper.GetStringSlice("columns")
 	config.OllamaAPIKey = viper.GetString("ollama_api_key")
 	config.OllamaAPIURL = viper.GetString("ollama_api_url")
+	config.OllamaModelsDir = viper.GetString("ollama_models_dir")
 	config.LMStudioFilePaths = viper.GetString("lm_studio_file_paths")
 	config.LogLevel = viper.GetString("log_level")
 	config.LogFilePath = viper.GetString("log_file_path")
