+++ config/config_test.go
@@ -61,6 +61,7 @@ func TestLoadConfig(t *testing.T) {
 					Columns:           []string{"Name", "Size"},
 					OllamaAPIKey:      "testkey",
 					OllamaAPIURL:      "http://testurl",
+					OllamaModelsDir:   GetOllamaModelDir(),
 					LMStudioFilePaths: "/test/path",
 					LogLevel:          "debug",
 					LogFilePath:       "/test/path/gollama.log",
@@ -75,6 +76,7 @@ func TestLoadConfig(t *testing.T) {
 				Columns:           []string{"Name", "Size"},
 				OllamaAPIKey:      "testkey",
 				OllamaAPIURL:      "http://testurl",
+				OllamaModelsDir:   GetOllamaModelDir(),
 				LMStudioFilePaths: "/test/path",
 				LogLevel:          "debug",
 				LogFilePath:       "/test/path/gollama.log",
@@ -97,6 +99,7 @@ func TestLoadConfig(t *testing.T) {
 				Columns:           []string{"Name", "Size", "Quant", "Family", "Modified", "ID"},
 				OllamaAPIKey:      "",
 				OllamaAPIURL:      "http://127.0.0.1:11434", // Default URL when no env vars are set
+				OllamaModelsDir:   GetOllamaModelDir(),
 				LMStudioFilePaths: "",
 				LogLevel:          "info",
 				SortOrder:         "modified",
@@ -119,6 +122,7 @@ func TestLoadConfig(t *testing.T) {
 				OllamaAPIKey:      "",
 				OllamaAPIURL:      "http://ollama.icu.lol",
 				LMStudioFilePaths: "",
+				OllamaModelsDir:   GetOllamaModelDir(),
 				LogLevel:          "info",
 				SortOrder:         "modified",
 				StripString:       "",
